# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Single sign-on (сервис единого входа)",
        "title": "SSO",
        "contact": {
            "name": "Alnovi",
            "url": "https://github.com/alnovi"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/alnovi/sso/refs/heads/master/LICENSE.md"
        },
        "version": "1.0.0"
    },
    "host": "http://localhost:8081",
    "paths": {
...
        "/oauth/authorize": {
            "get": {
                "description": "Форма аутентификации пользователя",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Форма аутентификации",
                "operationId": "OAuthForm",
                "parameters": [
                    {
                        "type": "string",
                        "description": "идентификатор клиента",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "тип запроса",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "адрес клиента (callback)",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "состояние",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Аутентификации пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Аутентификации пользователя",
                "operationId": "OAuthAuthorize",
                "parameters": [
                    {
                        "description": "Логин и пароль пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Authorize"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/oauth/certs": {
            "get": {
                "description": "Публичный ключ для проверки токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Публичный ключ",
                "operationId": "OAuthCerts",
                "responses": {
                    "200": {
                        "description": "Json web key",
                        "schema": {
                            "$ref": "#/definitions/response.JWK"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/oauth/forgot-password": {
            "get": {
                "description": "Форма получения токена для восстановления пароля",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Форма \"Забыли пароль\"",
                "operationId": "OAuthFormForgot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "идентификатор клиента",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "тип запроса",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "адрес клиента (callback)",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "состояние",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Отправка токена сброса пароля на почту пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Отправка токена сброса пароля",
                "operationId": "OAuthForgotPassword",
                "parameters": [
                    {
                        "description": "Логин пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ForgotPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/oauth/reset-password": {
            "get": {
                "description": "Форма для восстановления доступа",
                "consumes": [
                    "text/html"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Форма восстановления пароля",
                "operationId": "OAuthFormReset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "токен восстановления доступа",
                        "name": "hash",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "description": "Задать новый пароль пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Сброс пароля",
                "operationId": "OAuthResetPassword",
                "parameters": [
                    {
                        "description": "Новый пароль пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ResetPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.Authorize": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "minLength": 5,
                    "example": "name@example.com"
                },
                "password": {
                    "type": "string",
                    "maxLength": 24,
                    "minLength": 5,
                    "example": "qwerty"
                },
                "remember": {
                    "type": "boolean"
                }
            }
        },
...
        "request.ForgotPassword": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "minLength": 5,
                    "example": "name@example.com"
                }
            }
        },
        "request.ResetPassword": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 24,
                    "minLength": 5,
                    "example": "qwerty"
                },
                "token": {
                    "type": "string",
                    "example": "token-hash"
                }
            }
        },
...
        "response.JWK": {
            "type": "object",
            "properties": {
                "alg": {
                    "type": "string",
                    "example": "RS256"
                },
                "e": {
                    "type": "string",
                    "example": "AQAB"
                },
                "key_ops": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "verify"
                    ]
                },
                "kty": {
                    "type": "string",
                    "example": "RSA"
                },
                "n": {
                    "type": "string",
                    "example": "4TBai9lUV9qU0LqD37qpNTpJ1QuIn_2syDc9clGXKvf5lk6ESWaouvAT"
                },
                "use": {
                    "type": "string",
                    "example": "sig"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT-Access": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}